refactor:로그아웃 기능 수정 및 구조 변경-#34
